+++ glide3x/cvg/glide3/tests/test23.c
@@ -236,7 +236,7 @@ void main( int argc, char **argv) {
     return;
 }
 
-static unsigned long randx = 1;
+static unsigned int randx = 1;
 
 static unsigned int iRandom (unsigned int maxr)
 {
