+++ glide3x/cvg/glide3/tests/test36.c
@@ -35,7 +35,7 @@ void main( int argc, char **argv) {
                 iFailures = 0, iStatusAt = 0,
                 viewport[4];
         GrContext_t context;
-	unsigned long hWndMain = tlGethWnd();
+	unsigned int hWndMain = tlGethWnd();
 
         printf ( "Test36 - Glide resolution checker.  All output goes to the console.\n\n" );
         tlGetCH ();
