+++ swlibs/texus2/lib/texus.h
@@ -38,8 +38,8 @@ extern "C" {
 typedef unsigned char   FxU8;
 typedef unsigned short  FxU16;
 typedef short           FxI16;
-typedef unsigned long   FxU32;
-typedef long            FxI32;
+typedef unsigned int    FxU32;
+typedef int             FxI32;
 typedef int             FxBool;
 
 /*
